move newtype from legacy
a lil bit of json
