+++ web/src/engine/websites/_index.ts
@@ -299,6 +299,7 @@ export { default as MoonWitchInLove } from './MoonWitchInLove';
 export { default as MundoMangaKun } from './MundoMangaKun';
 export { default as Nekomik } from './Nekomik';
 export { default as NeoxScan } from './NeoxScan';
+export { default as NewType } from './NewType';
 export { default as Ngomik } from './Ngomik';
 export { default as NightComic } from './NightComic';
 export { default as Nightow } from './Nightow';
@@ -716,7 +717,6 @@ export { default as NeoProjectScans } from './legacy/NeoProjectScans';
 export { default as NetTruyen } from './legacy/NetTruyen';
 export { default as NeuManga } from './legacy/NeuManga';
 export { default as NewToki } from './legacy/NewToki';
-export { default as NewType } from './legacy/NewType';
 export { default as NhatTruyen } from './legacy/NhatTruyen';
 export { default as NHentai } from './legacy/NHentai';
 export { default as NicoNicoSeiga } from './legacy/NicoNicoSeiga';
